Moves `UIViewController` snapshot helpers to separate file
